chore(code): refactoring
